@@ -1,32 +1,40 @@
 <?php
 
-use Faker\Generator as Faker;
+namespace Database\Factories;
 
-/*
-|--------------------------------------------------------------------------
-| Model Factories
-|--------------------------------------------------------------------------
-|
-| This directory should contain each of the model factory definitions for
-| your application. Factories provide a convenient way to generate new
-| model instances for testing / seeding your application's database.
-|
-*/
+use App\Models\Character;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-$testServers = [
-    'Valefor',
-    'Ifrit',
-    'Ixion',
-    'Shiva',
-    'Bahamut',
-    'Yojimbo',
-    'Anima'
-];
-
-$factory->define(App\Character::class, function (Faker $faker) use ($testServers) {
-    return [
-        'lodestone_id' => $faker->randomNumber(),
-        'name' => $faker->firstName . ' ' . $faker->lastName,
-        'server' => $testServers[array_rand($testServers)]
+class CharacterFactory extends Factory
+{
+    private $testServers = [
+        'Valefor',
+        'Ifrit',
+        'Ixion',
+        'Shiva',
+        'Bahamut',
+        'Yojimbo',
+        'Anima'
     ];
-});
+
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Character::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'lodestone_id' => $this->faker->randomNumber(),
+            'name' => $this->faker->firstName . ' ' . $this->faker->lastName,
+            'server' => $this->faker->randomElement($this->testServers)
+        ];
+    }
+}
